fix: refactor code layout
