+++ example/apps/sqs/internal/pkg/cloud/aws.go
@@ -1,4 +1,4 @@
-package aws
+package cloud
 
 import (
 "github.com/aws/aws-sdk-go/aws"
