+++ example/apps/sqs/internal/pkg/cloud/config.go
@@ -1,4 +1,4 @@
-package aws
+package cloud
 
 import "os"
