+++ example/apps/sqs/internal/pkg/cloud/sqs.go
@@ -1,4 +1,4 @@
-package aws
+package cloud
 
 import (
 	"context"
@@ -6,13 +6,12 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/razorpay/devstack/example/apps/sqs/internal/pkg/cloud"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/sqs"
 )
 
-var _ cloud.MessageClient = SQS{}
+var _ MessageClient = SQS{}
 
 type SQS struct {
 	timeout time.Duration
@@ -93,7 +92,7 @@ func (s SQS) BindDLX(ctx context.Context, queueURL, dlxARN string) error {
 	return nil
 }
 
-func (s SQS) Send(ctx context.Context, req *cloud.SendRequest) (string, error) {
+func (s SQS) Send(ctx context.Context, req *SendRequest) (string, error) {
 	ctx, cancel := context.WithTimeout(ctx, s.timeout)
 	defer cancel()
 
@@ -117,7 +116,7 @@ func (s SQS) Send(ctx context.Context, req *cloud.SendRequest) (string, error) {
 	return *res.MessageId, nil
 }
 
-func (s SQS) Receive(ctx context.Context, queueURL string) (*cloud.Message, error) {
+func (s SQS) Receive(ctx context.Context, queueURL string) (*Message, error) {
 	// timeout = WaitTimeSeconds + 5
 	ctx, cancel := context.WithTimeout(ctx, time.Second*(20+5))
 	defer cancel()
@@ -141,7 +140,7 @@ func (s SQS) Receive(ctx context.Context, queueURL string) (*cloud.Message, erro
 		attrs[key] = *attr.StringValue
 	}
 
-	return &cloud.Message{
+	return &Message{
 		ID:            *res.Messages[0].MessageId,
 		ReceiptHandle: *res.Messages[0].ReceiptHandle,
 		Body:          *res.Messages[0].Body,
